fix: avoid relying exclusively on `CFStringGetCStringPtr()`
`CFStringGetCStringPtr()` fails and returns `nullptr` if the
string is not already available in memory as a string of the requested
encoding.

We have seen failures in mongosh CI with messages such as

```
MongoshInvalidInputError: Could not resolve certificate specification \
  'subject=tester@example.com': SecItemCopyMatching failed to \
  load certificates: (null)
```

which indicate that, for some reason, the switch to macOS 14 (and to
arm64) means that this case now occurs in our CI.

By avoiding `CFStringGetCStringPtr()` and instead using
`CFStringGetCString()`, we can likely avoid this issue.